refactor: PostTextArea 컴포넌트 분리
